Used this command:

    svn merge -r 1529475:HEAD https://svn.apache.org/repos/asf/roller/trunk .

git-svn-id: https://svn.apache.org/repos/asf/roller/branches/rome2@1546816 13f79535-47bb-0310-9956-ffa450edef68

+++ app/src/main/java/org/apache/roller/weblogger/webservices/adminprotocol/sdk/UserEntry.java
@@ -129,7 +129,7 @@ public class UserEntry extends Entry {
         // created (optional)
         Element createdElement = e.getChild(Tags.DATE_CREATED, Service.NAMESPACE);
         if (createdElement != null) {
-            setDateCreated(new Date(Long.valueOf(createdElement.getText()).longValue()));
+            setDateCreated(new Date(Long.valueOf(createdElement.getText())));
         }
         
         // enabled
